requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.7
django-crispy-forms==2.2
django-phonenumber-field==8.0.0
phonenumberslite==8.13.40
sqlparse==0.5.0
tzdata==2024.1
